Refactor across files
