+++ app/groceries.py
@@ -7,20 +7,8 @@
 
 import os
 
+from app.utils import to_usd
 
-def to_usd(my_price):
-    """
-    This is a docstring. It tells us what this function is about.
-    What its responsibilities are.
-    What the params are about 
-    What the datatypes the params are. 
-    What will this function will return 
-    Example of invoking the function. 
-    
-    Invoke like this: to_usd(9.9999)
-
-    """
-    return '${:,.2f}'.format(my_price)
 
 
 # checks to see if a products.csv file exists. If not, it uses the default
